move main page
